refactor: approval client into package
