+++ tests/tests_unit/approval_client/__init__.py
@@ -1,0 +1,3 @@
+from .client import ApprovalCogniteClient
+
+__all__ = ["ApprovalCogniteClient"]
